Hide DynamicConfigurationManager
