+++ agent/agent-sentinel/src/main/java/org/bithon/agent/sentinel/SentinelRuleManager.java
@@ -17,8 +17,7 @@
 package org.bithon.agent.sentinel;
 
 import org.bithon.agent.configuration.ConfigurationManager;
-import org.bithon.agent.controller.config.DynamicConfigurationManager;
-import org.bithon.agent.controller.config.IConfigurationChangedListener;
+import org.bithon.agent.configuration.IConfigurationChangedListener;
 import org.bithon.agent.sentinel.degrade.DegradingRuleDto;
 import org.bithon.agent.sentinel.expt.SentinelCommandException;
 import org.bithon.agent.sentinel.flow.FlowRuleDto;
@@ -73,7 +72,7 @@ public class SentinelRuleManager {
     }
 
     private SentinelRuleManager() {
-        DynamicConfigurationManager manager = DynamicConfigurationManager.getInstance();
+        ConfigurationManager manager = ConfigurationManager.getInstance();
         manager.addConfigurationChangeListener(new FlowRuleChangedListener());
         manager.addConfigurationChangeListener(new DegradingRuleChangedListener());
 
@@ -129,7 +128,7 @@ public class SentinelRuleManager {
 
             FlowRuleDto configRule = configRules.remove(key);
             if (configRule == null) {
-                // this rule in memory does not exist in current configuration
+                // this rule in memory does not exist in the current configuration
                 deleteRules.add(key);
             } else {
                 if (!configRule.equals(inMemoryRule)) {
@@ -203,7 +202,7 @@ public class SentinelRuleManager {
 
             DegradingRuleDto configRule = configRules.remove(key);
             if (configRule == null) {
-                // this rule in memory does not exist in current configuration
+                // this rule in memory does not exist in the current configuration
                 deleted.add(key);
             } else {
                 if (!configRule.equals(inMemoryRule)) {
